resources/archived/requirements.txt
pyservicebinding
numpy
pandas
seaborn
tensorflow
jupyter_contrib_nbextensions
jupyter_nbextensions_configurator
ipython-sql
sqlalchemy
psycopg2-binary